requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
charset-normalizer==3.2.0
discord.py==2.3.1
Django==4.2.3
frozenlist==1.4.0
gunicorn==21.2.0
idna==3.4
multidict==6.0.4
packaging==23.1
sqlparse==0.4.4
yarl==1.9.2
